refactor: clean up view chart view classes
